refactor(Makefile): Improve readability
This commit improves the readability of the Makefile by introducing
defines for common activities such as creating directories if they do
not exist, cloning git repositories and updating those git repositories
to the latest commits.

Furthermore, this commit also introduces a usability improvement by
parsing the Zephyr version directly from the zmk git repository. This is
done by parsing the embedded Dockerfile in the repository.